[GH-12976] Migrate 'components/admin_console/admin_settings.jsx' and associated tests to TypeScript (#4099)
* Migrate 'components/admin_console/admin_setting.jsx' to Typescript, update references and tests.

* Add onClick property to SaveButton.

* Remove duplicated onClick from SaveButton.
