+++ components/admin_console/cluster_settings.jsx
@@ -10,7 +10,7 @@ import * as Utils from 'utils/utils.jsx';
 import FormattedMarkdownMessage from 'components/formatted_markdown_message.jsx';
 import WarningIcon from 'components/widgets/icons/fa_warning_icon';
 
-import AdminSettings from './admin_settings.jsx';
+import AdminSettings from './admin_settings';
 import BooleanSetting from './boolean_setting.jsx';
 import ClusterTableContainer from './cluster_table_container.jsx';
 import SettingsGroup from './settings_group.jsx';
