+++ components/admin_console/database_settings.jsx
@@ -9,7 +9,7 @@ import {recycleDatabaseConnection} from 'actions/admin_actions.jsx';
 import * as Utils from 'utils/utils.jsx';
 import {t} from 'utils/i18n';
 
-import AdminSettings from './admin_settings.jsx';
+import AdminSettings from './admin_settings';
 import BooleanSetting from './boolean_setting.jsx';
 import RequestButton from './request_button/request_button.jsx';
 import SettingsGroup from './settings_group.jsx';
