+++ components/admin_console/message_export_settings.jsx
@@ -9,7 +9,7 @@ import {JobTypes} from 'utils/constants';
 import * as Utils from 'utils/utils.jsx';
 import {getSiteURL} from 'utils/url';
 
-import AdminSettings from './admin_settings.jsx';
+import AdminSettings from './admin_settings';
 import BooleanSetting from './boolean_setting.jsx';
 import DropdownSetting from './dropdown_setting.jsx';
 import JobsTable from './jobs';
