+++ components/admin_console/password_settings.jsx
@@ -8,7 +8,7 @@ import Constants from 'utils/constants';
 import * as Utils from 'utils/utils.jsx';
 import {t} from 'utils/i18n';
 
-import AdminSettings from './admin_settings.jsx';
+import AdminSettings from './admin_settings';
 import Setting from './setting';
 import SettingsGroup from './settings_group.jsx';
 import TextSetting from './text_setting.jsx';
