+++ components/admin_console/plugin_management/plugin_management.jsx
@@ -12,7 +12,7 @@ import LoadingScreen from 'components/loading_screen';
 import FormattedMarkdownMessage from 'components/formatted_markdown_message.jsx';
 import ConfirmModal from 'components/confirm_modal.jsx';
 
-import AdminSettings from '../admin_settings.jsx';
+import AdminSettings from '../admin_settings';
 import BooleanSetting from '../boolean_setting.jsx';
 import SettingsGroup from '../settings_group.jsx';
 import TextSetting from '../text_setting.jsx';
